Ieee1212ConfigRomLexer: add lexer of configuration ROM
In a commit 78c40c9f1022('Ieee1212ConfigRom: add configuration ROM parser
defined in IEEE 1212'), I committed processor of configuration ROM
according to IEEE 1212. However, the processor includes two parts which
are not tight-coupled.

This commit adds an alternative lexer implementation of the processor. For
backward compatility, the old processor is left as is for a while.