Ieee1212RootDirectoryParser: add parser of configuration ROM
In a commit 78c40c9f1022('Ieee1212ConfigRom: add configuration ROM parser
defined in IEEE 1212'), I committed processor of configuration ROM
according to IEEE 1212. However, the processor includes two parts which
are not tight-coupled.

This commit adds an alternative parser implementation of the processor,
to parse entries from lexer added in a former commit. For backward
compatility, the old processor is left as is for a while.