Ieee1394ConfigRomParser: add an alternative parser of configuration ROM in IEEE 1394
In former commits, a combination of lexer and parser of configuration ROM
in IEEE 1212 was added. They include no implementation for IEEE 1394.

This commit adds a parser for configuration ROM on IEEE 1394 bus.
Unfortunately, at present I have no oppoturnity to refer to IEEE 1394
specification. This commit is not fully compliant to the specification.
Old parser is left for a while.