ExtCtlSpace: add section parser and helper functions for commands in TCAT protocol extension
TC Applied Technology designed protocol extension as interfaces for
firmwares of its ASIC series. This protocol extension is called as
EAP; Extended Application Protocol.

This extension has 9 address sections for specific purposes. These sections
locate in a part of private space and offsets for the sections locate
in a fixed address space.

This commit adds a parser to detect the sections. Additionally, some helper
functions are added to read/write to each of the section.