ExtNewRouterSpace: add helper functions to handle router entries in New Router Space
In TCAT protocol extension, routing state between capture/playback ports.
channels in tx/rx packets, input/output of mixers locates in specifc
address region. A first quadlet in the region represents the number of
entries for routing information, the rest represents the entries, which
consists of one quadlet with four types of information; source,
destination and peak of signal.

This commit adds helper functions to handle the router entry.