ExtNewStreamConfigSpace: add helper functions to configure packet format
In TCAT protocol extension, packet format can be configurable in specific
address region of sixth section. The format is different from the one in
general protocol. When setting the configuration actually, programs should
initiate command to load from the space.

This commit adds helper functions to parse the configuration and to set
it.