FFMixerReg: add helper functions to handle registers for internal multiplexer
In a series of RME Fireface, internal multiplexer can be configured by
write transaction to specific regions. These regins are continuous area
in which one memory block is specific to one mixer element for one
output port. This block includes two parts; one is for source of physical
inputs and another is for source of stream inputs. One quadlet is used
to represent gain of each source for each mixer element.